Adding lots of missing documentation
Cleanup documentation for Collection


More documentation updates
