+++ src/DotNet.Plus/BasicType/EnumException.cs
@@ -2,6 +2,10 @@
 
 namespace DotNet.Plus.BasicType
 {
+    /// <summary>
+    /// This will be thrown when an operation is performed on an enum that was expected to be a [FLAGS]
+    /// enum.
+    /// </summary>
     public class NotFlagsEnumException : Exception
     {
         public NotFlagsEnumException(string message, Exception? innerException = null) :
@@ -10,7 +14,9 @@ namespace DotNet.Plus.BasicType
         }
     }
 
-
+    /// <summary>
+    /// This will be thrown when an object wasn't able to be converted to an Enum.
+    /// </summary>
     public class ConvertObjectToEnumException : Exception
     {
         public ConvertObjectToEnumException(string message, Exception? innerException = null) :
@@ -19,6 +25,10 @@ namespace DotNet.Plus.BasicType
         }
     }
 
+    /// <summary>
+    /// This will be thrown when an object wasn't able to be converted to an Enum with more detailed
+    /// information.
+    /// </summary>
     public class ConvertObjectToEnumException<TEnum> : ConvertObjectToEnumException
         where TEnum : Enum
     {
