+++ src/DotNet.Plus/Core/Operation.cs
@@ -1,9 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace DotNet.Plus.Core
 {
+    /// <summary>
+    /// Helpers to wrap try/catch operations.  This is a more functionality approach to try/catch management
+    /// and helps in unit testing failures.
+    /// </summary>
     public static class Operation
     {
         /// <summary>
