+++ src/DotNet.Plus/Endian/Endianness.cs
@@ -13,8 +13,7 @@ namespace DotNet.Plus.Endian
     {
         /// <summary>
         /// A big-endian system stores the most significant byte at the smallest memory address and the least significant byte at the largest.
-        /// </summary>
-        /// <example>
+        /// <code>
         /// Take the number 0x12345678
         ///
         ///   Pointer/  Big   
@@ -24,13 +23,13 @@ namespace DotNet.Plus.Endian
         ///     a+1:    0x34  
         ///     a+2:    0x56  
         ///     a+3:    0x78  
-        /// </example>
+        /// </code>
+        /// </summary>
         Big,
 
         /// <summary>
         /// A little-endian system stores the least-significant byte at the smallest address and the most significant byte at the largest.
-        /// </summary>
-        /// <example>
+        /// <code>
         /// Take the number 0x12345678
         ///
         ///   Pointer/  Little
@@ -40,16 +39,16 @@ namespace DotNet.Plus.Endian
         ///     a+1:     0x56
         ///     a+2:     0x34
         ///     a+3:     0x12
-        /// </example>
+        /// </code>
+        /// </summary>
         Little
     }
-    
+
     /// <summary>
     /// Endianness is the ordering or sequencing of bytes that represent numeric values.  It is expressed as big endian or little endian.
     /// A big-endian system stores the most significant byte at the smallest memory address and the least significant byte at the largest.
     /// A little-endian system, in contrast, stores the least-significant byte at the smallest address.
-    /// </summary>
-    /// <example>
+    /// <code>
     /// Take the number 0x12345678
     ///
     ///   Pointer/  Big     Little
@@ -59,7 +58,8 @@ namespace DotNet.Plus.Endian
     ///     a+1:    0x34     0x56
     ///     a+2:    0x56     0x34
     ///     a+3:    0x78     0x12
-    /// </example>
+    /// </code>
+    /// </summary>
     public static partial class Endianness
     {
         /// <summary>
