+++ src/DotNet.Plus/Pattern/Disposable.cs
@@ -1,7 +1,11 @@
 ﻿using System;
+using DotNet.Plus.Core;
 
 namespace DotNet.Plus.Pattern
 {
+    /// <summary>
+    /// Adds extensions to IDisposable such as for allowing easy dispose
+    /// </summary>
     public static class Disposable
     {
         /// <summary>
@@ -9,12 +13,7 @@ namespace DotNet.Plus.Pattern
         /// and silently handles any exceptions.
         /// </summary>
         /// <param name="instance">The instance to dispose</param>
-        public static void TryDispose(this IDisposable instance)
-        {
-            try {
-                instance.Dispose();
-            }
-            catch { /* ignored */ }
-        }
+        public static void TryDispose(this IDisposable instance) =>
+            Operation.TryCatch(instance.Dispose);
     }
 }
